+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   task to improve readability
 - removed extra the from notifcation of change task and add a transfer form
 - add simple side navigation to the statistics page
+- the Director of Child Services contact information is now in the External
+  contacts area, not the Local authority information area.
 
 ### Fixed